feat!: NCS 2.4 integration (#1)
* lib: fs_port: fix compilation error

`tmp` does not exist, changed it to `cursor`.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* lib: remove compilation warnings

This commit removes compilation warnings.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* lib: nrf_softsim: set ctx to NULL to fix re-init

Without this fix, a sequence of cfun=4 (or equivalent)
and cfun=1 would cause a freeze during the first APDU
due to the context not being properly reset.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* softsim: update integration with NCS

This commit updates the integration with NCS.

* Enforce Kconfig dependencies
* Enforce certain properties through build asserts
* Add software UICC selection to auto init
* Add debug Kconfig for pragma message
* Add a Kconfig overlay for all the required Kconfigs
* Updated manifest to align with NCS v2.4

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* softsim: README: add guide

This commit updates the README to add a guide on how to build an
application that uses Onomondo SoftSIM. It also contains notes regarding possible problems or errors and other useful information.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* lib: nrf_softsim: wrap callback in auto init

This commit moves the `on_modem_lib_init` callback
inside the `CONFIG_SOFTSIM_AUTO_INIT` conditional
compilation.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* Clean up lingering stuff from old commit

* remove internal ref.

---------

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>
Co-authored-by: Peter S. Bornerup <pb@onomondo.com>