feat: Add support for dynamic template (#4)
* softsim: add support for dynamic template

This commit adds support for dynamic template profile.
It builds a `template.hex` from `template.bin` based on the
address of the NVS storage partition as calculated by the
Partition Manager.
Added a Kconfig option for bundling the `template.hex` into
the final `merged.hex` of the application.
Removed build assert for fixed NVS storage partition base
address.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* lib: add provision check for file system

Previously the provision check was checking only
the PSA, now it checks also the file system.
This fixes a bug where the device would appear as
provisioned due to unchanged PSA, but being not provisioned
due to the NVS storage partition containing a fresh template
profile unprovisioned.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* lib: fs_port: change partition label

`storage_partition` is a partition that
changes based on the project configuration
and can be different things.
This commit changes the partition label to the
exact one (`nvs_storage`) to prevent pointing to the
wrong partition in certain cases.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

* softsim: fix typo

This commit fixes a typo.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>

---------

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>