fix: improve support for Thingy91
This commit addresses an issue where the samples do not compile out of
the box for the Thingy91 due to an overflow of the TFM partition.
Starting from the factory static partition layout, resize the `tfm`
partition to be large enough to satisfy the new size requirements
introduced by SoftSIM and update the other related addresses and sizes
accordingly.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>